+++ R/reporting.R
@@ -1,4 +1,10 @@
-#' @params corrplot_text_size Numeric, for the size of text label (variable names) for the coorplot diagram.
+#' @title Create a report from an STGrid object.
+#' @description
+#' Create a report from an STGrid object.
+#' @param corrplot_text_size Numeric, for the size of text label (variable names) for the coorplot diagram.
+#' @examples
+#' st_report(exp_1, spatial_image_params=list(ncol=4, features=feat_names(exp_1)[1:10]))
+#'
 st_report <- function(st_grid=NULL,
                       feat_list=NULL,
                       title="Spatial transcriptomics report",
@@ -19,11 +25,12 @@ st_report <- function(st_grid=NULL,
                       image_height=3,
                       #rmd_path = paste0(system.file(package = "mypackage"), "/rmd/template.Rmd"),
                       rmd_path=file.path("/Users/puthier/Documents/git/project_dev/STarlight",
-                                         "inst/rmd/basic_informations.Rmd"),
+                                         "inst/rmd/sample.Rmd"),
 
                       corrplot_text_size=0.5,
                       hc_tree_params=list(class_nb = 5,
                                           offset=5),
+                      plot_rip_k_params=list(ncol=4),
                       spatial_image_params=list(ncol=4,
                                                 features=NULL),
                       cmp_counts_st_params=list(fill_color="#7845FF",
@@ -50,7 +57,8 @@ st_report <- function(st_grid=NULL,
                   corrplot_text_size=corrplot_text_size,
                   hc_tree_params=hc_tree_params,
                   spatial_image_params=spatial_image_params,
-                  cmp_counts_st_params=cmp_counts_st_params),
+                  cmp_counts_st_params=cmp_counts_st_params,
+                  plot_rip_k_params=plot_rip_k_params),
     encoding     = 'UTF-8'
   )